test(components-react): move package test | tg | #2444
